refactor(profile): events module moved inside profile module, routers refactored
