+++ frontend/src/app/auth/pages/signin/signin.page.html
@@ -16,7 +16,7 @@
           <ion-item>
             <ion-label position="stacked">
               <b>Password</b>
-              <a [href]="'/recover'" float-right>Forgot password?</a>
+              <a routerLink="/recover" float-right>Forgot password?</a>
             </ion-label>
             <ion-input type="password" placeholder="Password" formControlName="password"></ion-input>
           </ion-item>
@@ -27,7 +27,7 @@
     </div>
 
     <div class="form-container" margin>
-      <p text-center class="auth-text">New to Actopus? <a color="secondary" [href]="'/signup'">Create an account.</a></p>
+      <p text-center class="auth-text">New to Actopus? <a color="secondary" routerLink="/signup">Create an account.</a></p>
     </div>
 
     <p text-center padding class="auth-text">Copyright 2018 @Actopus</p>
