+++ frontend/src/app/auth/pages/signup/signup.page.html
@@ -31,7 +31,7 @@
     </div>
 
     <div class="form-container" margin>
-      <p text-center class="auth-text">Already a user? <a color="secondary" [href]="'/signin'">Login here.</a></p>
+      <p text-center class="auth-text">Already a user? <a color="secondary" routerLink="/signin">Login here.</a></p>
     </div>
 
     <p text-center padding class="auth-text">Copyright 2018 @Actopus</p>
