+++ frontend/src/app/profile/events/components/event-card/event-card.component.spec.ts
