+++ frontend/src/app/profile/events/components/event-card/event-card.component.ts
@@ -1,5 +1,5 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {Event} from "../../../interfaces";
+import {Event} from "../../../../interfaces/index";
 
 @Component({
   selector: 'app-event-card',
