+++ frontend/src/app/profile/events/pages/event-create/event-create.page.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 
-import { EventService } from '../../../data/providers/event/event.service';
+import { EventService } from '../../../../data/providers/event/event.service';
 
 import {AlertController, LoadingController} from "@ionic/angular";
 import {HttpErrorResponse} from "@angular/common/http";
