+++ frontend/src/app/profile/events/pages/event/event.page.spec.ts
