+++ frontend/src/app/profile/events/pages/feed/feed.page.spec.ts
