+++ frontend/src/app/profile/events/pages/feed/feed.page.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
+import {ActivatedRoute, Router} from '@angular/router';
 
-import { Event } from '../../../interfaces';
+import { Event } from '../../../../interfaces/index';
 
-import { AuthService } from '../../../auth/providers/auth/auth.service';
-import { EventService } from '../../../data/providers/event/event.service';
+import { AuthService } from '../../../../auth/providers/auth/auth.service';
+import { EventService } from '../../../../data/providers/event/event.service';
 import {LoadingController} from "@ionic/angular";
 
 @Component({
@@ -16,7 +16,8 @@ export class FeedPage implements OnInit {
   events: Event[];
   private eventSub : any;
 
-  constructor(private router: Router, private authService: AuthService, private eventService: EventService, public loadingController: LoadingController) { }
+  constructor(private router: Router, private authService: AuthService, private eventService: EventService, public loadingController: LoadingController
+  ,private route : ActivatedRoute) { }
 
   ngOnInit() {
     this.presentLoading();
